Make scaling part of abstract BER type
Scaling factor is now a parameter of the `BER_finite` constructor
in `BER_float`. It is no longer considered as only present in
binary strings, but rather treated as a meaningful part of a
BER-encoded float.

Decision on whether the resulting `BER_float` is going to be
scaled is now made during the abstract IEEE->BER conversion.

All theorems remain the same, except the IEEE<->BER roundtrips,
which are now statements "forall scaling".